Make sure we don't try to remove any declarations twice
This could previously happen with removable methods on a removable class.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=179847512

+++ src/com/google/javascript/jscomp/ijs/PotentialDeclaration.java
@@ -65,7 +65,7 @@ class PotentialDeclaration {
   }
 
   static PotentialDeclaration fromMethod(Node functionNode, Scope scope) {
-    checkArgument(functionNode.isFunction());
+    checkArgument(ClassUtil.isClassMethod(functionNode));
     String name = ClassUtil.getPrototypeNameOfMethod(functionNode);
     return new PotentialDeclaration(name, functionNode.getParent(), functionNode, scope);
   }
@@ -89,11 +89,23 @@ class PotentialDeclaration {
     return NodeUtil.getBestJSDocInfo(lhs);
   }
 
+  private boolean isDetached() {
+    for (Node current = lhs; current != null; current = current.getParent()) {
+      if (current.isScript()) {
+        return false;
+      }
+    }
+    return true;
+  }
+
   /**
    * Remove this "potential declaration" completely.
    * Usually, this is because the same symbol has already been declared in this file.
    */
   void remove(AbstractCompiler compiler) {
+    if (isDetached()) {
+      return;
+    }
     Node statement = getStatement();
     NodeUtil.deleteNode(statement, compiler);
     statement.removeChildren();
